# documentation/api/swagger.yaml
openapi: 3.0.0
info:
  title: PulsePlus Game Platform API
  version: 1.0.0
  description: API documentation for the PulsePlus gaming and engagement platform
servers:
  - url: /api
    description: Main API server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    ChatGroup:
      type: object
      properties:
        sys_id:
          type: string
          format: uuid
        name:
          type: string
        creator_id:
          type: string
          format: uuid
        members:
          type: array
          items:
            type: string
            format: uuid
        type:
          type: string
          enum: [DIRECT, GROUP]
        created_at:
          type: string
          format: date-time
...
    SocialFeed:
      type: object
      properties:
        sys_id:
          type: string
          format: uuid
        content:
          type: string
        creator_id:
          type: string
          format: uuid
        attachments:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /social/feeds:
    post:
      tags:
        - Social
      summary: Create a new social feed post
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialFeed'
      responses:
        '201':
          description: Feed post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocialFeed'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /social/chat/groups:
    post:
      tags:
        - Social
      summary: Create a new chat group
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatGroup'
      responses:
        '201':
          description: Chat group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
security:
  - BearerAuth: [] 